requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-localflavor==3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.1
